# Get the gitignore cache working, consistently use fast JSON modules