--- a/tools/gitignore/gitignore.py
+++ b/tools/gitignore/gitignore.py
@@ -155,7 +155,10 @@ class PathFilter(object):
         self.literals_dir = defaultdict(dict)  # type: Dict[Optional[bytes], Dict[bytes, List[Tuple[bool, Pattern[bytes]]]]]
         self.patterns_file = []  # type: List[Tuple[Tuple[bool, Pattern[bytes]], List[Tuple[bool, Pattern[bytes]]]]]
         self.patterns_dir = []  # type: List[Tuple[Tuple[bool, Pattern[bytes]], List[Tuple[bool, Pattern[bytes]]]]]
-        self.cache = cache or {}  # type: MutableMapping[bytes, bool]
+
+        if cache is None:
+            cache = {}
+        self.cache = cache  # type: MutableMapping[bytes, bool]
 
         if extras is None:
             extras = []