--- a/tools/mypy.ini
+++ b/tools/mypy.ini
@@ -19,9 +19,6 @@ ignore_missing_imports = True
 [mypy-hypothesis.*]
 ignore_missing_imports = True
 
-[mypy-orjson.*]
-ignore_missing_imports = True
-
 [mypy-pytest.*]
 ignore_missing_imports = True