… tools/manifest/jsonlib.py