Some Sam Sneddon person broke the gitignore cache with e98eee5aec6 (as MutableMapping provides a default `__bool__`, effectively returning `bool(len(self))`), and this gets it working again.

This also slightly more consistently gets us using quicker JSON modules, and factors out a bunch of code for this which uses whichever we want. This provides a further benefit for those with them installed (`ujson` especially).

A few benchmarks (Py3.8, with a modified untracked file in the tree):

* master: 3.01s
* this branch: 2.82s
* this branch + orjson: 2.80s
* this branch + ujson: 2.17s
* this branch + orjson + ujson: 2.16s
* this branch + orjson + ujson patched for https://github.com/ultrajson/ultrajson/issues/429: 1.99s

(I don't remember why we set `caller_owns_obj=False` in `manifest.write`, given it exists for that reason, but with that set to `True` and with all the above we're down to 1.52s.)