@Hexcles 2020-10-05
Curious why we adding yet another fast JSON module fallback (orjson)?