@gsnedders 2020-10-05
> Curious why we adding yet another fast JSON module fallback (orjson)?

When we have it shared in a common place, the cost of adding it is pretty low. That said, for all that's been posted in many places about its performance its gains are… pretty minute. Happy to remove it, given that.